backend/requirements-render.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pulp>=2.7.0
numpy>=1.21.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
gunicorn==21.2.0